Move ModuleRecoveryAlert to its own folder.
